+++ src/app/components/timer-page/timer-page.component.ts
@@ -1,7 +1,6 @@
 import { Component, OnInit, ViewChild, Input } from '@angular/core';
 
 import { AnimationContainerGroupComponent } from '../animation-container-group/animation-container-group.component';
-import { AnimationContainerComponent } from '../animation-container/animation-container.component';
 import { TimerComponent, TimerDisplayInfo } from '../timer/timer.component';
 import { Timer, TimerState } from '../../class/timer';
 
@@ -13,14 +12,14 @@ import { Timer, TimerState } from '../../class/timer';
 export class TimerPageComponent implements OnInit {
 
   @ViewChild('containerGroup') containerGroup: AnimationContainerGroupComponent;
-  @ViewChild('timerAddingContainer') timerAddingContainer: AnimationContainerComponent;
-  @ViewChild('timersContainer') timersContainer: AnimationContainerComponent;
-
-  @ViewChild('timersContainerGroup') timersContainerGroup: AnimationContainerGroupComponent;
   @ViewChild('inputComponent') inputComponent: TimerComponent;
 
   timers: Timer[] = [];
-  pageNum: number;
+
+  pageInfo: { state: number, lastIndex: number } = {
+    state: 1,
+    lastIndex: 1
+  };
 
   constructor() { }
 
@@ -28,7 +27,9 @@ export class TimerPageComponent implements OnInit {
   }
 
   ngAfterViewInit() {
-    this.showContainerPage(1);
+    setTimeout(() => {
+      this.showContainerPage(1);
+    }, 0);
   }
 
   getTimerInfos(state: TimerState): TimerDisplayInfo {
@@ -41,30 +42,63 @@ export class TimerPageComponent implements OnInit {
       textFlashing: state === TimerState.Paused || state === TimerState.TimesUp,
     };
   }
-  getBtnVisibility() {
+  getVisibilities() {
 
     let totalSeconds = this.inputComponent?.getInputTotalSeconds() ?? 0;
 
     return {
-      add: this.pageNum === 1 && totalSeconds > 0,
-      cancel: this.pageNum === 1 && this.timers.length > 0,
-      addTimer: this.pageNum === 2,
-      delete: this.pageNum === 2 && this.timers.length > 0,
+      showSlicks: this.pageInfo.state === 2,
+      add: this.pageInfo.state === 1 && totalSeconds > 0,
+      cancel: this.pageInfo.state === 1 && this.timers.length > 0,
+      addTimer: this.pageInfo.state === 2,
+      delete: this.pageInfo.state === 2 && this.timers.length > 0,
     };
   }
 
-  showContainerPage(num: number) {
+  showContainerPage(code: number) {
 
-    if (num === 1) {
+    let toIndex = 0;
+    let tabLength = this.containerGroup.tabLength;
+    let currentIndex = this.containerGroup.currentIndex;
+    let lastIndex = this.pageInfo.lastIndex;
+
+    switch (code) {
+      case 1: // add timer page
+        lastIndex = currentIndex > 0 ? currentIndex : 1;
+        toIndex = 0;
+        break;
+      case 2: // cancal timer page
+        toIndex = lastIndex;
+        break;
+      case 3: // add timer
+        toIndex = tabLength;
+        break;
+      case 4: // delete timer
+        toIndex = currentIndex - 1;
+        tabLength -= 1;
+        break;
+    }
+
+    if (tabLength == 1) {
+      toIndex = 0;
+    }
+    if (toIndex == 0) {
       this.inputComponent.clearInput();
-      this.containerGroup.switchToItem(this.timerAddingContainer);
     }
-    else if (num === 2) {
-      this.containerGroup.switchToItem(this.timersContainer);
-    }
-    this.pageNum = num;
+
+    this.containerGroup.switchToIndex(toIndex, true);
+    this.pageInfo = {
+      state: toIndex <= 0 ? 1 : 2,
+      lastIndex: lastIndex
+    };
   }
 
+  switchToAddingTimer() {
+    this.showContainerPage(1);
+  }
+  cancalAddingTimer() {
+    this.showContainerPage(2);
+  }
   addTimer() {
 
     let timer = new Timer();
@@ -75,25 +109,18 @@ export class TimerPageComponent implements OnInit {
     this.timers.push(timer);
 
     setTimeout(() => {
-      this.timersContainerGroup.switchToLast();
-      this.showContainerPage(2);
+      this.showContainerPage(3);
     }, 0);
   }
   deleteTimer() {
 
-    let currentIndex = this.timersContainerGroup.currentIndex;
+    let currentIndex = this.containerGroup.currentIndex - 1;
     if (currentIndex >= 0) {
 
-      if (this.timers.length <= 1) {
-        this.showContainerPage(1);
-      } else {
-        this.showContainerPage(2);
-      }
-      this.timersContainerGroup.switchToIndex(currentIndex - 1);
+      this.showContainerPage(4);
 
       this.timers[currentIndex].unsubscribe();
       this.timers.splice(currentIndex, 1);
-
     }
   }