+++ CHANGELOG.md
@@ -32,7 +32,7 @@
 # CHANGE LOG
 
 - 20210408
-  1. animation-container.component: improve switching animation
+  1. animation-container.component: improve switching mechanism and animation
 - 20210407
   1. animation-container.component: change switchItem to switch index
   2. animation-container.component: imporve slick style